Delete redundant core code from prev API
